Update ThermostatModeSwitch.hms
The methode .HSSID() is only working for datapoints, not for channels and not for devices. While e.g. 
var x = channel.HSSID();
causes an runtime error, is 
if( "CLIMATECONTROL_RT_TRANSCEIVER" == channel.HssType() )
simply evaluated always to TRUE. In the old version the inner block is evaluateted for every channel in the CCU!